feat(eventspage): create different aspect pages for both mobile and desktop
