+++ package.json
@@ -19,6 +19,9 @@
         "prepare": "husky install"
     },
     "dependencies": {
+        "@emotion/react": "^11.11.1",
+        "@emotion/styled": "^11.11.0",
+        "@mui/material": "^5.14.16",
         "@uidotdev/usehooks": "^2.4.0",
         "autoprefixer": "^10.4.16",
         "clsx": "^1.2.1",
